homework7/arrow function, module
